refactor: restructure project files to adhere to Clean Architecture
